+++ CHANGES.md
@@ -5,9 +5,12 @@
 - Support using async error processor and async spec processor
 ([pull #57][pull_57]).
 - Fix auto-tag support for nesting blueprint ([pull #58][pull_58]).
+- Support set the url prefix of the OpenAPI blueprint with the 
+`openapi_blueprint_url_prefix` argument ([pull #64][pull_64]).
 
 [pull_57]: https://github.com/greyli/apiflask/pull/57
 [pull_58]: https://github.com/greyli/apiflask/pull/58
+[pull_64]: https://github.com/greyli/apiflask/pull/64
 
 
 ## Version 0.6.3